+++ packages/components-js/tests/vrt/specs/accordion.vrt.ts
@@ -10,15 +10,15 @@ import {
   testOptions,
 } from '../helpers';
 
-describe('Panel', () => {
+describe('Accordion', () => {
   it('should have no visual regression', async () => {
     const vrt = getVisualRegressionTester();
     expect(
       await vrt.test(
-        'panel',
+        'accordion',
         async () => {
           await vrt.goTo('/#text'); // to ensure fonts are already loaded before js is initialized
-          await vrt.goTo('/#panel');
+          await vrt.goTo('/#accordion');
         },
         testOptions
       )
@@ -28,18 +28,18 @@ describe('Panel', () => {
   it('should have no visual regression for :hover + :focus-visible', async () => {
     const vrt = getVisualRegressionStatesTester();
     expect(
-      await vrt.test('panel-states', async () => {
+      await vrt.test('accordion-states', async () => {
         const page = await vrt.getPage();
-        const hoverElementSelector = 'p-panel >>> p-headline';
-        const focusElementSelector = 'p-panel >>> button';
+        const hoverElementSelector = 'p-accordion >>> p-headline';
+        const focusElementSelector = 'p-accordion >>> button';
 
         const getElementsMarkup: GetThemedMarkup = (theme) => `
-          <p-panel theme="${theme}" heading="Some heading">
+          <p-accordion theme="${theme}" heading="Some heading">
             <p>
              Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat,
              sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
             </p>
-          </p-panel>`;
+          </p-accordion>`;
 
         await setContentWithDesignSystem(page, getThemedBodyMarkup(getElementsMarkup));