+++ packages/components-react/tests/vrt/specs/accordion.vrt.ts
@@ -1,14 +1,14 @@
 import { getVisualRegressionTester, testOptions } from '../helpers';
 
-describe('Panel', () => {
+describe('Accordion', () => {
   it('should have no visual regression', async () => {
     const vrt = getVisualRegressionTester();
     expect(
       await vrt.test(
-        'panel',
+        'accordion',
         async () => {
           await vrt.goTo('/text'); // to ensure fonts are already loaded before js is initialized
-          await vrt.goTo('/panel');
+          await vrt.goTo('/accordion');
         },
         testOptions
       )