+++ packages/components/src/components/content/accordion/accordion-utils.ts
@@ -1,0 +1,3 @@
+const ACCORDION_HEADER_SIZE = ['small', 'medium'] as const;
+export type AccordionHeaderSize = typeof ACCORDION_HEADER_SIZE[number];
+export type AccordionChangeEvent = { open: boolean };